Not finished yet. See details.
This is intended to be a color selector tool. So far, you will see a
square on the left, which is really a coordinate system where x values
represent shades of red, and y values represend shades of blue. On the
right, there is a rectangle which shows an array of green values, whose
red and blue values are determined by where the cursor is on the box at
the left. If you click on the box, your selection of red and blue will
be saved. Click outside of the box, and you will unselect the saved
values.

What I will do with this is have a similar selecting scheme for the
rectangle on the right, and then have the rgb values shown as numbers
on the screen so that they can be used as a reference, similar to the
color selector feature in the IDE.